chore: reorganize unit tests for addons-v5 package
