+++ packages/addons-v5/test/unit/commands/addons/attach.unit.test.js
