+++ packages/addons-v5/test/unit/commands/addons/destroy.unit.test.js
