+++ packages/addons-v5/test/unit/commands/addons/detach.unit.test.js
