+++ packages/addons-v5/test/unit/commands/addons/docs.unit.test.js
