+++ packages/addons-v5/test/unit/lib/util.unit.test.js
@@ -1,6 +1,6 @@
 'use strict'
 
-let util = require('../../lib/util')
+let util = require('../../../lib/util')
 let expect = require('chai').expect
 
 describe('util.formatPrice', function () {
